# Conflicts:
#	sormas-api/src/main/resources/enum.properties
#	sormas-api/src/main/resources/enum_de-CH.properties
#	sormas-api/src/main/resources/enum_de-DE.properties
#	sormas-api/src/main/resources/enum_en-AF.properties
#	sormas-api/src/main/resources/enum_es-CU.properties
#	sormas-api/src/main/resources/enum_es-EC.properties
#	sormas-api/src/main/resources/enum_es-ES.properties
#	sormas-api/src/main/resources/enum_fa-AF.properties
#	sormas-api/src/main/resources/enum_fi-FI.properties
#	sormas-api/src/main/resources/enum_fil-PH.properties
#	sormas-api/src/main/resources/enum_fj-FJ.properties
#	sormas-api/src/main/resources/enum_fr-CH.properties
#	sormas-api/src/main/resources/enum_fr-FR.properties
#	sormas-api/src/main/resources/enum_haw-US.properties
#	sormas-api/src/main/resources/enum_hi-IN.properties
#	sormas-api/src/main/resources/enum_it-CH.properties
#	sormas-api/src/main/resources/enum_it-IT.properties
#	sormas-api/src/main/resources/enum_ja-JP.properties
#	sormas-api/src/main/resources/enum_ks-PK.properties
#	sormas-api/src/main/resources/enum_ku-TR.properties
#	sormas-api/src/main/resources/enum_la-LA.properties
#	sormas-api/src/main/resources/enum_mt-MT.properties
#	sormas-api/src/main/resources/enum_nl-NL.properties
#	sormas-api/src/main/resources/enum_no-NO.properties
#	sormas-api/src/main/resources/enum_pl-PL.properties
#	sormas-api/src/main/resources/enum_ps-AF.properties
#	sormas-api/src/main/resources/enum_pt-PT.properties
#	sormas-api/src/main/resources/enum_ro-RO.properties
#	sormas-api/src/main/resources/enum_ru-RU.properties
#	sormas-api/src/main/resources/enum_sv-SE.properties
#	sormas-api/src/main/resources/enum_sw-KE.properties
#	sormas-api/src/main/resources/enum_tr-TR.properties
#	sormas-api/src/main/resources/enum_uk-UA.properties
#	sormas-api/src/main/resources/enum_ur-PK.properties
#	sormas-api/src/main/resources/enum_zh-CN.properties
#	sormas-api/src/main/resources/enum_zu-ZA.properties
#	sormas-ui/src/main/java/de/symeda/sormas/ui/utils/FieldHelper.java

+++ sormas-ui/src/main/java/de/symeda/sormas/ui/events/EventsView.java
@@ -17,11 +17,6 @@
  *******************************************************************************/
 package de.symeda.sormas.ui.events;
 
-import java.util.Date;
-import java.util.HashMap;
-
-import org.vaadin.hene.popupbutton.PopupButton;
-
 import com.vaadin.icons.VaadinIcons;
 import com.vaadin.navigator.ViewChangeListener.ViewChangeEvent;
 import com.vaadin.server.StreamResource;
@@ -37,9 +32,10 @@ import com.vaadin.ui.VerticalLayout;
 import com.vaadin.ui.Window;
 import com.vaadin.ui.themes.ValoTheme;
 import com.vaadin.v7.ui.ComboBox;
-
+import com.vaadin.v7.ui.OptionGroup;
 import de.symeda.sormas.api.EntityRelevanceStatus;
 import de.symeda.sormas.api.FacadeProvider;
+import de.symeda.sormas.api.action.ActionStatus;
 import de.symeda.sormas.api.event.EventCriteria;
 import de.symeda.sormas.api.event.EventDto;
 import de.symeda.sormas.api.event.EventExportDto;
@@ -53,6 +49,7 @@ import de.symeda.sormas.api.user.UserRight;
 import de.symeda.sormas.api.utils.DateHelper;
 import de.symeda.sormas.ui.ControllerProvider;
 import de.symeda.sormas.ui.SearchSpecificLayout;
+import de.symeda.sormas.ui.SormasUI;
 import de.symeda.sormas.ui.UserProvider;
 import de.symeda.sormas.ui.ViewModelProviders;
 import de.symeda.sormas.ui.utils.AbstractView;
@@ -60,11 +57,15 @@ import de.symeda.sormas.ui.utils.ButtonHelper;
 import de.symeda.sormas.ui.utils.CssStyles;
 import de.symeda.sormas.ui.utils.DateFormatHelper;
 import de.symeda.sormas.ui.utils.DownloadUtil;
+import de.symeda.sormas.ui.utils.FilteredGrid;
 import de.symeda.sormas.ui.utils.GridExportStreamResource;
 import de.symeda.sormas.ui.utils.LayoutUtil;
 import de.symeda.sormas.ui.utils.MenuBarHelper;
 import de.symeda.sormas.ui.utils.VaadinUiUtil;
-import de.symeda.sormas.ui.utils.ViewConfiguration;
+import org.vaadin.hene.popupbutton.PopupButton;
+
+import java.util.Date;
+import java.util.HashMap;
 
 public class EventsView extends AbstractView {
 
@@ -73,9 +74,9 @@ public class EventsView extends AbstractView {
 	public static final String VIEW_NAME = "events";
 
 	private EventCriteria criteria;
-	private ViewConfiguration viewConfiguration;
+	private EventsViewConfiguration viewConfiguration;
 
-	private EventGrid grid;
+	private FilteredGrid<?, ?> grid;
 	private Button createButton;
 	private HashMap<Button, String> statusButtons;
 	private Button activeStatusButton;
@@ -87,23 +88,29 @@ public class EventsView extends AbstractView {
 
 	private VerticalLayout gridLayout;
 
-	private String originalViewTitle;
-
 	// Bulk operations
 	private MenuBar bulkOperationsDropdown;
 
 	public EventsView() {
 		super(VIEW_NAME);
 
-		originalViewTitle = getViewTitleLabel().getValue();
+		viewConfiguration = ViewModelProviders.of(getClass()).get(EventsViewConfiguration.class);
+		if (viewConfiguration.getViewType() == null) {
+			viewConfiguration.setViewType(EventsViewType.DEFAULT);
+		}
 
-		viewConfiguration = ViewModelProviders.of(getClass()).get(ViewConfiguration.class);
 		criteria = ViewModelProviders.of(EventsView.class).get(EventCriteria.class);
 		if (criteria.getRelevanceStatus() == null) {
 			criteria.relevanceStatus(EntityRelevanceStatus.ACTIVE);
 		}
 
-		grid = new EventGrid(criteria, getClass());
+		if (isDefaultViewType()) {
+			grid = new EventGrid(criteria, getClass());
+			grid.getDataProvider().addDataProviderListener(e -> updateStatusButtons());
+		} else {
+			grid = new EventActionsGrid(criteria, getClass());
+			grid.getDataProvider().addDataProviderListener(e -> updateStatusButtons());
+		}
 		gridLayout = new VerticalLayout();
 		gridLayout.addComponent(createFilterBar());
 		gridLayout.addComponent(createStatusFilterBar());
@@ -113,9 +120,26 @@ public class EventsView extends AbstractView {
 		gridLayout.setSizeFull();
 		gridLayout.setExpandRatio(grid, 1);
 		gridLayout.setStyleName("crud-main-layout");
-		grid.getDataProvider().addDataProviderListener(e -> updateStatusButtons());
 
 		addComponent(gridLayout);
+
+		OptionGroup eventsViewSwitcher = new OptionGroup();
+		eventsViewSwitcher.setId("eventsViewSwitcher");
+		CssStyles.style(eventsViewSwitcher, CssStyles.FORCE_CAPTION, ValoTheme.OPTIONGROUP_HORIZONTAL, CssStyles.OPTIONGROUP_HORIZONTAL_PRIMARY);
+		eventsViewSwitcher.addItem(EventsViewType.DEFAULT);
+		eventsViewSwitcher.setItemCaption(EventsViewType.DEFAULT, I18nProperties.getCaption(Captions.eventDefaultView));
+
+		eventsViewSwitcher.addItem(EventsViewType.ACTIONS);
+		eventsViewSwitcher.setItemCaption(EventsViewType.ACTIONS, I18nProperties.getCaption(Captions.eventActionsView));
+
+		eventsViewSwitcher.setValue(viewConfiguration.getViewType());
+		eventsViewSwitcher.addValueChangeListener(e -> {
+			EventsViewType viewType = (EventsViewType) e.getProperty().getValue();
+
+			viewConfiguration.setViewType(viewType);
+			SormasUI.get().getNavigator().navigateTo(EventsView.VIEW_NAME);
+		});
+		addHeaderComponent(eventsViewSwitcher);
 
 		if (UserProvider.getCurrent().hasUserRight(UserRight.EVENT_EXPORT)) {
 			VerticalLayout exportLayout = new VerticalLayout();
@@ -136,35 +160,37 @@ public class EventsView extends AbstractView {
 			}
 
 			{
-				StreamResource exportStreamResource = DownloadUtil.createCsvExportStreamResource(
-					EventExportDto.class,
-					null,
-					(Integer start, Integer max) -> FacadeProvider.getEventFacade().getExportList(grid.getCriteria(), start, max),
-					(propertyId, type) -> {
-						String caption = I18nProperties.findPrefixCaption(
-							propertyId,
-							EventExportDto.I18N_PREFIX,
-							EventIndexDto.I18N_PREFIX,
-							EventDto.I18N_PREFIX,
-							LocationDto.I18N_PREFIX);
-						if (Date.class.isAssignableFrom(type)) {
-							caption += " (" + DateFormatHelper.getDateFormatPattern() + ")";
-						}
-						return caption;
-					},
-					createFileNameWithCurrentDate("sormas_events_", ".csv"),
-					null);
-				addExportButton(
-					exportStreamResource,
-					exportPopupButton,
-					exportLayout,
-					VaadinIcons.FILE_TEXT,
-					Captions.exportDetailed,
-					Strings.infoDetailedExport);
+				if (isDefaultViewType()) {
+					StreamResource exportStreamResource = DownloadUtil.createCsvExportStreamResource(
+						EventExportDto.class,
+						null,
+						(Integer start, Integer max) -> FacadeProvider.getEventFacade().getExportList((EventCriteria) grid.getCriteria(), start, max),
+						(propertyId, type) -> {
+							String caption = I18nProperties.findPrefixCaption(
+								propertyId,
+								EventExportDto.I18N_PREFIX,
+								EventIndexDto.I18N_PREFIX,
+								EventDto.I18N_PREFIX,
+								LocationDto.I18N_PREFIX);
+							if (Date.class.isAssignableFrom(type)) {
+								caption += " (" + DateFormatHelper.getDateFormatPattern() + ")";
+							}
+							return caption;
+						},
+						createFileNameWithCurrentDate("sormas_events_", ".csv"),
+						null);
+					addExportButton(
+						exportStreamResource,
+						exportPopupButton,
+						exportLayout,
+						VaadinIcons.FILE_TEXT,
+						Captions.exportDetailed,
+						Strings.infoDetailedExport);
+				}
 			}
 		}
 
-		if (UserProvider.getCurrent().hasUserRight(UserRight.PERFORM_BULK_OPERATIONS)) {
+		if (UserProvider.getCurrent().hasUserRight(UserRight.PERFORM_BULK_OPERATIONS) && isDefaultViewType()) {
 			Button btnEnterBulkEditMode = ButtonHelper.createIconButton(Captions.actionEnterBulkEditMode, VaadinIcons.CHECK_SQUARE_O, null);
 			btnEnterBulkEditMode.setVisible(!viewConfiguration.isInEagerMode());
 
@@ -181,8 +207,8 @@ public class EventsView extends AbstractView {
 				viewConfiguration.setInEagerMode(true);
 				btnEnterBulkEditMode.setVisible(false);
 				btnLeaveBulkEditMode.setVisible(true);
-				grid.setEagerDataProvider();
-				grid.reload();
+				((EventGrid) grid).setEagerDataProvider();
+				((EventGrid) grid).reload();
 			});
 			btnLeaveBulkEditMode.addClickListener(e -> {
 				bulkOperationsDropdown.setVisible(false);
@@ -203,12 +229,18 @@ public class EventsView extends AbstractView {
 			addHeaderComponent(createButton);
 		}
 
-		Button searchSpecificEventButton = ButtonHelper.createIconButton(
-			Captions.eventSearchSpecificEvent,
-			VaadinIcons.SEARCH,
-			e -> buildAndOpenSearchSpecificEventWindow(),
-			ValoTheme.BUTTON_PRIMARY);
-		addHeaderComponent(searchSpecificEventButton);
+		if (isDefaultViewType()) {
+			Button searchSpecificEventButton = ButtonHelper.createIconButton(
+				Captions.eventSearchSpecificEvent,
+				VaadinIcons.SEARCH,
+				e -> buildAndOpenSearchSpecificEventWindow(),
+				ValoTheme.BUTTON_PRIMARY);
+			addHeaderComponent(searchSpecificEventButton);
+		}
+	}
+
+	private boolean isDefaultViewType() {
+		return viewConfiguration.getViewType() == EventsViewType.DEFAULT;
 	}
 
 	public HorizontalLayout createFilterBar() {
@@ -217,7 +249,7 @@ public class EventsView extends AbstractView {
 		filterLayout.setMargin(false);
 		filterLayout.setSizeUndefined();
 
-		filterForm = new EventsFilterForm();
+		filterForm = new EventsFilterForm(isDefaultViewType(), isDefaultViewType());
 		filterForm.addValueChangeListener(e -> {
 			if (!filterForm.hasFilter()) {
 				navigateTo(null);
@@ -275,28 +307,54 @@ public class EventsView extends AbstractView {
 
 		statusButtons = new HashMap<>();
 
-		Button statusAll = ButtonHelper.createButton(Captions.all, e -> {
-			criteria.eventStatus(null);
-			navigateTo(criteria);
-		}, ValoTheme.BUTTON_BORDERLESS, CssStyles.BUTTON_FILTER);
-		statusAll.setCaptionAsHtml(true);
-
-		statusFilterLayout.addComponent(statusAll);
-
-		statusButtons.put(statusAll, I18nProperties.getCaption(Captions.all));
-		activeStatusButton = statusAll;
-
-		for (EventStatus status : EventStatus.values()) {
-			Button statusButton = ButtonHelper.createButtonWithCaption("status-" + status, status.toString(), e -> {
-				criteria.eventStatus(status);
+		if (isDefaultViewType()) {
+			Button statusAll = ButtonHelper.createButton(Captions.all, e -> {
+				criteria.setEventStatus(null);
 				navigateTo(criteria);
-			}, ValoTheme.BUTTON_BORDERLESS, CssStyles.BUTTON_FILTER, CssStyles.BUTTON_FILTER_LIGHT);
-			statusButton.setCaptionAsHtml(true);
-			statusButton.setData(status);
+			}, ValoTheme.BUTTON_BORDERLESS, CssStyles.BUTTON_FILTER);
+			statusAll.setCaptionAsHtml(true);
 
-			statusFilterLayout.addComponent(statusButton);
+			statusFilterLayout.addComponent(statusAll);
 
-			statusButtons.put(statusButton, status.toString());
+			statusButtons.put(statusAll, I18nProperties.getCaption(Captions.all));
+			activeStatusButton = statusAll;
+
+			for (EventStatus status : EventStatus.values()) {
+				Button statusButton = ButtonHelper.createButtonWithCaption("status-" + status, status.toString(), e -> {
+					criteria.setEventStatus(status);
+					navigateTo(criteria);
+				}, ValoTheme.BUTTON_BORDERLESS, CssStyles.BUTTON_FILTER, CssStyles.BUTTON_FILTER_LIGHT);
+				statusButton.setCaptionAsHtml(true);
+				statusButton.setData(status);
+
+				statusFilterLayout.addComponent(statusButton);
+
+				statusButtons.put(statusButton, status.toString());
+			}
+		} else {
+			Button statusAll = ButtonHelper.createButton(Captions.all, e -> {
+				criteria.setActionStatus(null);
+				navigateTo(criteria);
+			}, ValoTheme.BUTTON_BORDERLESS, CssStyles.BUTTON_FILTER);
+			statusAll.setCaptionAsHtml(true);
+
+			statusFilterLayout.addComponent(statusAll);
+
+			statusButtons.put(statusAll, I18nProperties.getCaption(Captions.all));
+			activeStatusButton = statusAll;
+
+			for (ActionStatus status : ActionStatus.values()) {
+				Button statusButton = ButtonHelper.createButtonWithCaption("status-" + status, status.toString(), e -> {
+					criteria.actionStatus(status);
+					navigateTo(criteria);
+				}, ValoTheme.BUTTON_BORDERLESS, CssStyles.BUTTON_FILTER, CssStyles.BUTTON_FILTER_LIGHT);
+				statusButton.setCaptionAsHtml(true);
+				statusButton.setData(status);
+
+				statusFilterLayout.addComponent(statusButton);
+
+				statusButtons.put(statusButton, status.toString());
+			}
 		}
 
 		HorizontalLayout actionButtonsLayout = new HorizontalLayout();
@@ -332,24 +390,25 @@ public class EventsView extends AbstractView {
 			}
 
 			// Bulk operation dropdown
-			if (UserProvider.getCurrent().hasUserRight(UserRight.PERFORM_BULK_OPERATIONS)) {
+			if (UserProvider.getCurrent().hasUserRight(UserRight.PERFORM_BULK_OPERATIONS) && isDefaultViewType()) {
+				EventGrid eventGrid = (EventGrid) grid;
 				bulkOperationsDropdown = MenuBarHelper.createDropDown(
 					Captions.bulkActions,
 					new MenuBarHelper.MenuBarItem(I18nProperties.getCaption(Captions.bulkEdit), VaadinIcons.ELLIPSIS_H, selectedItem -> {
-						ControllerProvider.getEventController().showBulkEventDataEditComponent(grid.asMultiSelect().getSelectedItems());
+						ControllerProvider.getEventController().showBulkEventDataEditComponent(eventGrid.asMultiSelect().getSelectedItems());
 					}),
 					new MenuBarHelper.MenuBarItem(I18nProperties.getCaption(Captions.bulkDelete), VaadinIcons.TRASH, selectedItem -> {
 						ControllerProvider.getEventController()
-							.deleteAllSelectedItems(grid.asMultiSelect().getSelectedItems(), () -> navigateTo(criteria));
+							.deleteAllSelectedItems(eventGrid.asMultiSelect().getSelectedItems(), () -> navigateTo(criteria));
 					}),
 					new MenuBarHelper.MenuBarItem(I18nProperties.getCaption(Captions.actionArchive), VaadinIcons.ARCHIVE, selectedItem -> {
 						ControllerProvider.getEventController()
-							.archiveAllSelectedItems(grid.asMultiSelect().getSelectedItems(), () -> navigateTo(criteria));
-					}),
+							.archiveAllSelectedItems(eventGrid.asMultiSelect().getSelectedItems(), () -> navigateTo(criteria));
+					}, EntityRelevanceStatus.ACTIVE.equals(criteria.getRelevanceStatus())),
 					new MenuBarHelper.MenuBarItem(I18nProperties.getCaption(Captions.actionDearchive), VaadinIcons.ARCHIVE, selectedItem -> {
 						ControllerProvider.getEventController()
-							.dearchiveAllSelectedItems(grid.asMultiSelect().getSelectedItems(), () -> navigateTo(criteria));
-					}, false));
+							.dearchiveAllSelectedItems(eventGrid.asMultiSelect().getSelectedItems(), () -> navigateTo(criteria));
+					}, EntityRelevanceStatus.ARCHIVED.equals(criteria.getRelevanceStatus())));
 
 				bulkOperationsDropdown.setVisible(viewConfiguration.isInEagerMode());
 				actionButtonsLayout.addComponent(bulkOperationsDropdown);
@@ -369,8 +428,17 @@ public class EventsView extends AbstractView {
 			params = params.substring(1);
 			criteria.fromUrlParams(params);
 		}
+
+		if (viewConfiguration.isInEagerMode() && isDefaultViewType()) {
+			((EventGrid) grid).setEagerDataProvider();
+		}
+
 		updateFilterComponents();
-		grid.reload();
+		if (isDefaultViewType()) {
+			((EventGrid) grid).reload();
+		} else {
+			((EventActionsGrid) grid).reload();
+		}
 	}
 
 	public void updateFilterComponents() {
@@ -393,7 +461,7 @@ public class EventsView extends AbstractView {
 		statusButtons.keySet().forEach(b -> {
 			CssStyles.style(b, CssStyles.BUTTON_FILTER_LIGHT);
 			b.setCaption(statusButtons.get(b));
-			if (b.getData() == criteria.getEventStatus()) {
+			if (b.getData() == (isDefaultViewType() ? criteria.getEventStatus() : criteria.getActionStatus())) {
 				activeStatusButton = b;
 			}
 		});
